+++ packages/ui/src/components/widgets/SwapWidgetRenderer.tsx
@@ -470,7 +470,7 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
     required: gasTopUpRequired,
     amount: _gasTopUpAmount,
     amountUsd: _gasTopUpAmountUsd
-  } = useGasTopUpRequired(toChain, toToken, recipient)
+  } = useGasTopUpRequired(toChain, fromChain, toToken, recipient)
 
   const quoteParameters =
     fromToken && toToken
@@ -533,7 +533,9 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
           debouncedOutputAmountValue.length > 0 &&
           Number(debouncedOutputAmountValue) !== 0)) &&
       fromToken !== undefined &&
-      toToken !== undefined
+      toToken !== undefined &&
+      !transactionModalOpen &&
+      !depositAddressModalOpen
   )
 
   const {
@@ -574,10 +576,9 @@ const SwapWidgetRenderer: FC<SwapWidgetRendererProps> = ({
   const invalidateQuoteQuery = useCallback(() => {
     queryClient.invalidateQueries({ queryKey: quoteQueryKey })
   }, [queryClient, quoteQueryKey])
-
-  let error = _quoteData || isFetchingQuote ? null : quoteError
+  let error =
+    _quoteData || (isFetchingQuote && quoteFetchingEnabled) ? null : quoteError
   let quote = error ? undefined : _quoteData
-
   const gasTopUpAmount = quote?.details?.currencyGasTopup?.amount
     ? BigInt(quote?.details?.currencyGasTopup?.amount)
     : _gasTopUpAmount